[feat] #11 my page fragment로 변경
