+++ core-designsystem/src/main/res/color/selector_menu_color.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/main_3" android:state_checked="true" />
+    <item android:color="@color/gray_400" android:state_checked="false" />
+</selector>
