+++ core-ui/src/main/java/com/sopt/now/coreui/base/BindingFragment.kt
@@ -10,7 +10,7 @@ import androidx.viewbinding.ViewBinding
 abstract class BindingFragment<T : ViewBinding>(private val inflater: (LayoutInflater) -> T) :
     Fragment() {
     private var _binding: T? = null
-    private val binding
+    val binding
         get() = requireNotNull(_binding) { "${this::class.java.simpleName}에서 에러가 발생했습니다." }
